+++ libs/nx_monitoring/src/nx/monitoring/activity_monitor.h
@@ -38,7 +38,7 @@ public:
          * suitable to be shown to the user. */
         QString partitions;
 
-        friend uint qHash(const Hdd& hdd) {
+        friend size_t qHash(const Hdd& hdd) {
             return ::qHash(hdd.id) ^ qHash(hdd.name) ^ qHash(hdd.partitions);
         }
     };
