+++ vms/client/nx_vms_client_desktop/src/nx/vms/client/desktop/event_search/right_panel_globals.h
@@ -102,14 +102,14 @@ struct EventCategory
 
 void registerQmlType();
 
-inline uint qHash(RightPanel::TimeSelection source)
+inline size_t qHash(RightPanel::TimeSelection source)
 {
-    return uint(source);
+    return size_t(source);
 }
 
-inline uint qHash(RightPanel::CameraSelection source)
+inline size_t qHash(RightPanel::CameraSelection source)
 {
-    return uint(source);
+    return size_t(source);
 }
 
 inline uint qHash(RightPanel::SystemSelection source)
