+++ vms/client/nx_vms_client_desktop/src/nx/vms/client/desktop/shaders/media_output_shader_program.h
@@ -82,6 +82,6 @@ private:
     const Key m_key;
 };
 
-uint qHash(const MediaOutputShaderProgram::Key& key, uint seed = 0);
+size_t qHash(const MediaOutputShaderProgram::Key& key, uint seed = 0);
 
 } // namespace nx::vms::client::desktop
