+++ vms/client/nx_vms_client_desktop/src/ui/common/text_pixmap_cache.cpp
@@ -135,7 +135,7 @@ public:
         {
         }
 
-        friend uint qHash(const Key& key)
+        friend size_t qHash(const Key& key)
         {
             using ::qHash;
             return
